+++ piecrust/CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Change event handling to emit `piecrust-uplink::Event`
+- Change `emit` export to include topic
 - Remove `Into<PathBuf>` bound in `VM::new`
 - Rename `host_debug` export to `hdebug` [#222]
 
@@ -23,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Remove `Event` struct
 - Remove `__heap_base` requirement from contracts
 
 ## [0.5.0] - 2023-06-07